coding ...

+++ registry.py
@@ -31,13 +31,14 @@ class Registry():
             self._services[image['Id']] = image
 
             # liang: dump the images, need to move to cache abstraction in the future.
-            logger.info('dumping %s' % image['Id'])
             if image['Id'] == u'3d3b49d80014e2df3434f282586b3bb2cff0f7b5f58a3e63d9229c48085a53a8':
                 continue
-            raw = self._cli.get_image(image['Id'])
-            tar = open(conf['image_dir']+image['Id']+'.tar', 'w')
-            tar.write(raw.data)
-            tar.close()
+            if not os.path.exists(conf['image_dir']+image['Id']+'.tar'):
+                logger.info('dumping %s' % image['Id'])
+                raw = self._cli.get_image(image['Id'])
+                tar = open(conf['image_dir']+image['Id']+'.tar', 'w')
+                tar.write(raw.data)
+                tar.close()
         pass
 
 
@@ -53,7 +54,11 @@ class Registry():
         pass
 
 
-    def start_service(self):
+    def start_service(self, sid, cmd):
+        #self._cli.create_container(image=sid, command=cmd)
+        container = self._cli.create_container(image='busybox:latest', command=cmd)
+        print container
+        self._cli.start(sid)
         pass
 
 
@@ -74,6 +79,7 @@ class Registry():
 def test():
     cli = Client(conf['base_url'])
     reg = Registry(cli)
+    #reg.start_service(u'8c2e06607696bd4afb3d03b687e361cc43cf8ec1a4a725bc96e39f05ba97dd55', '/bin/sleep 30')
     pass
